# gateway-service/docs/swagger.yaml
basePath: /
definitions:
  main.FailResponse:
    properties:
      status:
        type: string
    type: object
  main.LoginSuccessResponse:
    properties:
      status:
        type: string
      token:
        type: string
    type: object
  main.MobilePaymentRequest:
    properties:
      account_id:
        type: string
      amount:
        type: number
      callback_url:
        type: string
      channel:
        type: string
      narration:
        type: string
      payment_method:
        type: string
      phone_number:
        type: string
    type: object
  main.PaymentRequest:
    properties:
      amount:
        type: number
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      location:
        type: string
      payment_method:
        type: string
      phone:
        type: string
      reason:
        type: string
      username:
        type: string
    type: object
  main.SuccessResponse:
    properties:
      user:
        $ref: '#/definitions/main.User'
    type: object
  main.User:
    properties:
      email:
        type: string
      location:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  main.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: 54.145.134.156:8083
info:
  contact:
    email: keithkadima@gmail.com
    name: API Support
  description: This is a payment gateway service that integrates with the authentication
    and payments services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Payment Gateway API
  version: "0.1"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with user details and token
          schema:
            $ref: '#/definitions/main.LoginSuccessResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/main.FailResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/main.FailResponse'
      summary: Login a user
      tags:
      - auth
  /payments/initiate:
    post:
      consumes:
      - application/json
      description: Initiate a payment for a user
      parameters:
      - description: Payment Request
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/main.PaymentRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Initiate a payment
      tags:
      - payments
  /payments/send-to-mobile:
    post:
      consumes:
      - application/json
      description: Send money to a mobile number via the Payd API
      parameters:
      - description: Mobile Payment Request
        in: body
        name: mobilePayment
        required: true
        schema:
          $ref: '#/definitions/main.MobilePaymentRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Send money to a mobile number
      tags:
      - payments
  /payments/status/{id}:
    get:
      description: Get the status of a payment by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get payment status
      tags:
      - payments
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided details
      parameters:
      - description: User Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/main.SuccessResponse'
        "401":
          description: Registration failed
          schema:
            $ref: '#/definitions/main.FailResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/main.FailResponse'
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"